improved multiple lists name rendering

+++ src/Lists.js
@@ -1,5 +1,6 @@
 const List = require("./List");
-const ListToChooseElement = require("./ListToChooseElement");
+
+const ListToChooseElement = require("./elements/ListToChooseElement");
 const myLocalStorage = require("./MyLocalStorage");
 const listsElement = document.getElementById("lists");
 const addListFormElement = document.getElementById("add-list-form");
@@ -27,28 +28,44 @@ class Lists {
     myLocalStorage.setItem("selectedListId", this.selectedListId);
   }
   addList(name) {
-    console.log(this.selectedListId);
-    let id = new Date().getTime();
-    let list = { id: id, name: name, itemsUndone: [], itemsCompleted: [] };
-    this.lists.set(list.id, list);
-    this.elements.set(list.id, new ListToChooseElement(list));
-    Lists.listsEl.append(this.elements.get(id));
+    let alreadyExistingList = null;
+    this.lists.forEach((list) => {
+      if (list.name === name) {
+        alreadyExistingList = list;
+      }
+    });
+    if (!alreadyExistingList) {
+      let id = new Date().getTime();
+      let list = { id: id, name: name, itemsUndone: [], itemsCompleted: [] };
+      this.lists.set(list.id, list);
+      this.elements.set(list.id, new ListToChooseElement(list));
+      Lists.listsEl.append(this.elements.get(id));
+      Lists.addListFormEl.reset();
+      Lists.inputListEl.value = "";
+      myLocalStorage.setItem(list.id, list);
+      this.selectList(id);
+    } else {
+      this.selectList(alreadyExistingList.id);
+    }
     Lists.addListFormEl.reset();
     Lists.inputListEl.value = "";
-    myLocalStorage.setItem(list.id, list);
-    this.selectList(id);
   }
   deleteList(id) {
+    console.log(id);
     this.lists.delete(id);
     this.elements.get(id).remove();
     this.elements.delete(id);
-    myLocalStorage.remove(id);
+    myLocalStorage.removeItem(id);
     this.selectedListId = "";
     ///select another list
     this.save();
   }
   selectList(id) {
+    if (this.selectedListId) {
+      this.elements.get(this.selectedListId).classList.remove("hidden");
+    }
     this.selectedListId = id;
+    this.elements.get(this.selectedListId).classList.add("hidden");
     this.save();
   }
   static get listsEl() {
